# Add JMH benchmarks for issue 1075