Before 26f5d5f8 (benchmarks run atop 98ff897):
```
Benchmark                                        Mode  Cnt     Score     Error   Units
EmitterBenchmark.emitScalar                      avgt    3     0.311 ±   0.009   us/op
EmitterBenchmark.emitScalar:·gc.alloc.rate       avgt    3  6697.479 ± 420.474  MB/sec
EmitterBenchmark.emitScalar:·gc.alloc.rate.norm  avgt    3   256.000 ±   0.001    B/op
EmitterBenchmark.emitScalar:·gc.count            avgt    3   108.000            counts
EmitterBenchmark.emitScalar:·gc.time             avgt    3    62.000                ms
```

After 26f5d5f8 (benchmarks run atop 8a3063b):
```
Benchmark                                        Mode  Cnt     Score     Error   Units
EmitterBenchmark.emitScalar                      avgt    3     0.220 ±   0.006   us/op
EmitterBenchmark.emitScalar:·gc.alloc.rate       avgt    3  4713.283 ± 303.238  MB/sec
EmitterBenchmark.emitScalar:·gc.alloc.rate.norm  avgt    3   128.000 ±   0.001    B/op
EmitterBenchmark.emitScalar:·gc.count            avgt    3   116.000            counts
EmitterBenchmark.emitScalar:·gc.time             avgt    3    63.000                ms
```

The benchmarks were run on a macbook, so cpu scaling makes the timing values a bit chaotic and I wouldn't put too much trust in the ~30% speedup shown. The important piece is `gc.alloc.rate.norm` which is the garbage collected bytes per `emitScalar` invocation, which cut in half.